apps/api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
strawberry-graphql[fastapi]>=0.219.0
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
redis>=5.0.1
boto3>=1.34.34
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-multipart>=0.0.6
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

